chore(ui.config): move all popper arrow to config
