+++ src/runtime/ui.config/popper.ts
@@ -1,5 +1,5 @@
-export default {
-	base: 'before:w-2 before:h-2',
+export const arrow = {
+	base: 'invisible before:visible before:block before:rotate-45 before:z-[-1] before:w-2 before:h-2',
 	ring: 'before:ring-1 before:ring-gray-200 dark:before:ring-gray-800',
 	rounded: 'before:rounded-sm',
 	background: 'before:bg-gray-200 dark:before:bg-gray-800',
